v2.1.0 #28 Updates for latest WDL
*DEVS* Please update to latest WDL and repatch.  swell-ini.cpp is the only file that needs to be patched now.  I will eventually integrate the patch into SWS so no more patching is necessary.

+++ Misc/Adam.cpp
@@ -2895,7 +2895,7 @@ void AWSplitXFadeLeft(COMMAND_T* t)
 	{
         dItemLength = GetMediaItemInfo_Value(items.Get()[i], "D_LENGTH");
         dItemStart = GetMediaItemInfo_Value(items.Get()[i], "D_POSITION");
-        item1group = GetMediaItemInfo_Value(items.Get()[i], "I_GROUPID");
+        item1group = (int)GetMediaItemInfo_Value(items.Get()[i], "I_GROUPID");
 
         if (cursorPos > dItemStart && cursorPos < (dItemStart + dItemLength))
         {
